+++ test/integrations/kissmetrics.js
@@ -75,21 +75,21 @@ describe('KISSmetrics', function () {
       assert(window._kmq.push.calledWith(['record', 'Viewed Name Page', {}]));
     });
 
-    it('should not track named pages when the option is off', function () {
+    it('should track named pages with categories', function () {
+      kissmetrics.page('Category', 'Name');
+      assert(window._kmq.push.calledWith(['record', 'Viewed Category Name Page', {}]));
+    });
+
+    it('should track categorized pages by default', function () {
+      kissmetrics.page('Category', 'Name');
+      assert(window._kmq.push.calledWith(['record', 'Viewed Category Page', {}]));
+    });
+
+    it('should not track a named or categorized page when the option is off', function () {
       kissmetrics.options.trackNamedPages = false;
+      kissmetrics.options.trackCategorizedPages = false;
       kissmetrics.page(null, 'Name');
-      assert(!window._kmq.push.called);
-    });
-
-    it('should track sectioned pages by default', function () {
-      kissmetrics.page('Section', 'Name');
-      assert(window._kmq.push.calledWith(['record', 'Viewed Section Name Page', {}]));
-    });
-
-    it('should not track sectioned pages when the option is off', function () {
-      kissmetrics.options.trackNamedPages = false;
-      kissmetrics.options.trackSectionedPages = false;
-      kissmetrics.page('Section', 'Name');
+      kissmetrics.page('Category', 'Name');
       assert(!window._kmq.push.called);
     });
   });